splited profile file into as dashboard and profile and renamed patient.py as dashboard.py
