+++ phr/__init__.py
@@ -506,7 +506,7 @@ def getProfileImage(data):
 @frappe.whitelist(allow_guest=True)
 def getEmergencyDetails(data):
 	data = json.loads(data)
-	from templates.pages.profile import get_user_details
+	from templates.pages.dashboard import get_user_details
 	user_details = get_user_details(data.get('profile_id'))
 	if user_details.get('error'):
 		return user_details
