--- phr/templates/pages/patient.py
@@ -1,2 +1,0 @@
-
-""" this file has splited in two files form_generator.py and utils.py"""
