+++ phr/templates/pages/todo.py
@@ -3,7 +3,7 @@ import datetime
 from erpnext.setup.doctype.sms_settings.sms_settings import send_sms
 from frappe.utils.email_lib import sendmail
 import json
-from phr.templates.pages.profile import search_profile_data_from_solr
+from phr.templates.pages.dashboard import search_profile_data_from_solr
 from phr.templates.pages.utils import send_phrs_mail,get_formatted_date_time,get_sms_template
